requirements.txt
absl-py==2.0.0
astroid==3.3.10
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.3
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
deepface==0.0.92
dill==0.4.0
dlib==19.24.2
dnspython==2.4.2
easydict==1.13
easyocr==1.7.2
eventlet==0.33.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
idna==3.4
imageio==2.37.0
imutils==0.5.4
intel-openmp==2021.4.0
isort==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.23.0
libclang==18.1.1
mariadb==1.1.8
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.7
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mrz==0.6.2
mtcnn==0.1.1
mysqlclient==2.2.3
namex==0.0.8
networkx==3.3
ninja==1.11.1.3
numpy==1.26.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.12.0
packaging==23.2
pandas==2.2.2
PassportEye==2.2.2
pdfminer==20191125
Pillow==10.1.0
platformdirs==4.3.8
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycodestyle==2.13.0
pycparser==2.21
pycryptodome==3.22.0
pydub==0.25.1
pyflakes==3.3.2
Pygments==2.18.0
pylint==3.3.7
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.6.3
python-dateutil==2.8.2
python-engineio==4.8.0
python-Levenshtein==0.23.0
python-socketio==5.10.0
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.5.2
requests==2.31.0
retina-face==0.0.17
rich==13.7.1
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
shapely==2.0.7
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.32
srt==3.5.3
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tesseract==0.1.3
tf_keras==2.16.0
threadpoolctl==3.6.0
tifffile==2025.1.10
tomlkit==0.13.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.3.116
ultralytics-thop==2.0.14
urllib3==2.0.7
vosk==0.3.45
Werkzeug==3.0.0
wrapt==1.16.0
wsproto==1.2.0
